Modified as requested by review.
