+++ enterprise/micronaut/src/org/netbeans/modules/micronaut/MicronautConfigUtilities.java
@@ -18,36 +18,18 @@
  */
 package org.netbeans.modules.micronaut;
 
-import java.awt.Toolkit;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.EnumSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.function.Consumer;
 import java.util.regex.Pattern;
-import javax.lang.model.element.ElementKind;
-import javax.lang.model.element.ExecutableElement;
-import javax.lang.model.element.TypeElement;
-import javax.lang.model.util.ElementFilter;
 import javax.swing.text.Document;
 import org.netbeans.api.editor.document.EditorDocumentUtils;
 import org.netbeans.api.editor.document.LineDocument;
 import org.netbeans.api.editor.document.LineDocumentUtils;
-import org.netbeans.api.editor.mimelookup.MimeRegistration;
-import org.netbeans.api.java.source.ClasspathInfo;
-import org.netbeans.api.java.source.ElementHandle;
-import org.netbeans.api.java.source.JavaSource;
-import org.netbeans.api.java.source.ui.ElementOpen;
-import org.netbeans.api.progress.BaseProgressUtils;
 import org.netbeans.api.project.FileOwnerQuery;
 import org.netbeans.api.project.Project;
-import org.netbeans.lib.editor.hyperlink.spi.HyperlinkLocation;
-import org.netbeans.lib.editor.hyperlink.spi.HyperlinkProviderExt;
-import org.netbeans.lib.editor.hyperlink.spi.HyperlinkType;
 import org.netbeans.modules.csl.api.StructureItem;
 import org.netbeans.modules.csl.api.StructureScanner;
 import org.netbeans.modules.csl.core.Language;
@@ -57,10 +39,10 @@ import org.netbeans.modules.parsing.api.ParserManager;
 import org.netbeans.modules.parsing.api.ResultIterator;
 import org.netbeans.modules.parsing.api.Source;
 import org.netbeans.modules.parsing.api.UserTask;
+import org.netbeans.modules.parsing.spi.ParseException;
 import org.netbeans.modules.parsing.spi.Parser;
 import org.openide.filesystems.FileObject;
 import org.openide.util.Exceptions;
-import org.openide.util.NbBundle;
 import org.springframework.boot.configurationmetadata.ConfigurationMetadataGroup;
 import org.springframework.boot.configurationmetadata.ConfigurationMetadataProperty;
 import org.springframework.boot.configurationmetadata.ConfigurationMetadataSource;
@@ -69,83 +51,16 @@ import org.springframework.boot.configurationmetadata.ConfigurationMetadataSourc
  *
  * @author Dusan Balek
  */
-@MimeRegistration(mimeType = "text/x-yaml", service = HyperlinkProviderExt.class)
-public class MicronautConfigHyperlinkProvider implements HyperlinkProviderExt {
+public class MicronautConfigUtilities {
 
-    @Override
-    public Set<HyperlinkType> getSupportedHyperlinkTypes() {
-        return EnumSet.of(HyperlinkType.GO_TO_DECLARATION);
-    }
-
-    @Override
-    public boolean isHyperlinkPoint(Document doc, int offset, HyperlinkType type) {
-        return getHyperlinkSpan(doc, offset, type) != null;
-    }
-
-    @Override
-    public int[] getHyperlinkSpan(Document doc, int offset, HyperlinkType type) {
-        int[] span = new int[2];
-        ConfigurationMetadataProperty property = resolve(doc, offset, span, null);
-        return property != null ? span : null;
-    }
-
-    @Override
-    @NbBundle.Messages("LBL_GoToDeclaration=Go to Declaration")
-    public void performClickAction(Document doc, int offset, HyperlinkType type) {
-        AtomicBoolean cancel = new AtomicBoolean();
-        BaseProgressUtils.runOffEventDispatchThread(() -> {
-            List<ConfigurationMetadataSource> sources = new ArrayList<>();
-            ConfigurationMetadataProperty property = resolve(doc, offset, null, sources);
-            if (property != null && !sources.isEmpty()) {
-                ClasspathInfo cpInfo = ClasspathInfo.create(doc);
-                ElementHandle handle = getElementHandle(cpInfo, sources.get(0).getType(), property.getName(), cancel);
-                if (handle == null || !ElementOpen.open(cpInfo, handle)) {
-                    Toolkit.getDefaultToolkit().beep();
-                }
-            }
-        }, Bundle.LBL_GoToDeclaration(), cancel, false);
-    }
-
-    @Override
-    public String getTooltipText(Document doc, int offset, HyperlinkType type) {
-        ConfigurationMetadataProperty property = resolve(doc, offset, null, null);
-        if (property != null) {
-            StringBuilder sb = new StringBuilder("<html><body>");
-            sb.append("<b>").append(property.getId().replace(".", /* ZERO WIDTH SPACE */".&#x200B;")).append("</b>");
-            String propertyType = property.getType();
-            if (propertyType != null) {
-                sb.append("<pre>").append(propertyType).append("</pre>");
-            }
-            return sb.toString();
-        }
-        return null;
-    }
-
-    @Override
-    public CompletableFuture<HyperlinkLocation> getHyperlinkLocation(Document doc, int offset, HyperlinkType type) {
-        final AtomicBoolean cancel = new AtomicBoolean();
-        List<ConfigurationMetadataSource> sources = new ArrayList<>();
-        ConfigurationMetadataProperty property = resolve(doc, offset, null, sources);
-        if (property != null && !sources.isEmpty()) {
-            ClasspathInfo cpInfo = ClasspathInfo.create(doc);
-            String typeName = sources.get(0).getType();
-            ElementHandle handle = getElementHandle(cpInfo, typeName, property.getName(), cancel);
-            if (handle != null) {
-                return ElementOpen.getLocation(cpInfo, handle, typeName.replace('.', '/') + ".class");
-            }
-        }
-        return CompletableFuture.completedFuture(null);
-    }
-
-    static ConfigurationMetadataProperty resolve(Document doc, int offset, int[] span, List<ConfigurationMetadataSource> sources) {
+    public static ConfigurationMetadataProperty resolveProperty(Document doc, int offset, int[] span, List<ConfigurationMetadataSource> sources) {
         LineDocument lineDocument = LineDocumentUtils.as(doc, LineDocument.class);
         if (lineDocument != null) {
             FileObject fo = EditorDocumentUtils.getFileObject(doc);
             if (fo != null && "application.yml".equalsIgnoreCase(fo.getNameExt())) {
                 Project project = FileOwnerQuery.getOwner(fo);
                 if (project != null) {
-                    final MicronautConfigProperties configProperties = project.getLookup().lookup(MicronautConfigProperties.class);
-                    if (configProperties != null) {
+                    if (MicronautConfigProperties.hasConfigMetadata(project)) {
                         try {
                             int lineStart = LineDocumentUtils.getLineStart(lineDocument, offset);
                             String text = lineDocument.getText(lineStart, offset - lineStart);
@@ -173,8 +88,7 @@ public class MicronautConfigHyperlinkProvider implements HyperlinkProviderExt {
                                                             }
                                                             if (start <= offset && offset <= end && item.getName().equals(lineDocument.getText(start, end - start))) {
                                                                 String propertyName = getPropertyName(context);
-                                                                Map<String, ConfigurationMetadataGroup> groups = configProperties.getGroups();
-                                                                for (Map.Entry<String, ConfigurationMetadataGroup> groupEntry : configProperties.getGroups().entrySet()) {
+                                                                for (Map.Entry<String, ConfigurationMetadataGroup> groupEntry : MicronautConfigProperties.getGroups(project).entrySet()) {
                                                                     String groupKey = groupEntry.getKey();
                                                                     if (Pattern.matches(groupKey.replaceAll("\\.", "\\\\.").replaceAll("\\*", "\\\\w*") + ".*", propertyName)) {
                                                                         ConfigurationMetadataGroup group = groupEntry.getValue();
@@ -209,6 +123,44 @@ public class MicronautConfigHyperlinkProvider implements HyperlinkProviderExt {
         return null;
     }
 
+    public static void collectUsages(FileObject fo, String propertyName, Consumer<Usage> consumer) {
+        try {
+            ParserManager.parse(Collections.singleton(Source.create(fo)), new UserTask() {
+                public @Override void run(ResultIterator resultIterator) throws Exception {
+                    Parser.Result r = resultIterator.getParserResult();
+                    if (r instanceof ParserResult) {
+                        Language language = LanguageRegistry.getInstance().getLanguageByMimeType(resultIterator.getSnapshot().getMimeType());
+                        if (language != null) {
+                            StructureScanner scanner = language.getStructure();
+                            if (scanner != null) {
+                                find(fo, propertyName, scanner.scan((ParserResult) r), r.getSnapshot().getText(), consumer);
+                            }
+                        }
+                    }
+                }
+            });
+        } catch (ParseException ex) {
+            Exceptions.printStackTrace(ex);
+        }
+    }
+
+    private static void find(FileObject fo, String propertyName, List<? extends StructureItem> structures, CharSequence content, Consumer<Usage> consumer) {
+        int idx = propertyName.indexOf('.');
+        String name = idx < 0 ? propertyName : propertyName.substring(0, idx);
+        for (StructureItem structure : structures) {
+            if ("*".equals(name) || name.equals(structure.getName())) {
+                if (idx < 0) {
+                    int start = (int) structure.getPosition();
+                    int end = (int) structure.getEndPosition();
+                    String text = content.subSequence(start, end).toString();
+                    consumer.accept(new Usage(fo, start, end, text));
+                } else {
+                    find(fo, propertyName.substring(idx + 1), structure.getNestedItems(), content, consumer);
+                }
+            }
+        }
+    }
+
     private static List<StructureItem> getContext(List<? extends StructureItem> structure, int offset) {
         List<StructureItem> context = new ArrayList<>();
         loop: while (structure != null && !structure.isEmpty()) {
@@ -235,28 +187,34 @@ public class MicronautConfigHyperlinkProvider implements HyperlinkProviderExt {
         return sb.toString();
     }
 
-    private ElementHandle getElementHandle(ClasspathInfo cpInfo, String typeName, String propertyName, AtomicBoolean cancel) {
-        ElementHandle[] handle = new ElementHandle[1];
-        if (cpInfo != null) {
-            try {
-                JavaSource.create(cpInfo).runUserActionTask(controller -> {
-                    if (cancel != null && cancel.get()) {
-                        return;
-                    }
-                    handle[0] = ElementHandle.createTypeElementHandle(ElementKind.CLASS, typeName);
-                    TypeElement te = (TypeElement) handle[0].resolve(controller);
-                    if (te != null) {
-                        String name = "set" + propertyName.replaceAll("-", "");
-                        for (ExecutableElement executableElement : ElementFilter.methodsIn(te.getEnclosedElements())) {
-                            if (name.equalsIgnoreCase(executableElement.getSimpleName().toString())) {
-                                handle[0] = ElementHandle.create(executableElement);
-                                break;
-                            }
-                        }
-                    }
-                }, true);
-            } catch (IOException ex) {}
+    public static final class Usage {
+
+        private final FileObject fileObject;
+        private final int startOffset;
+        private final int endOffset;
+        private final String text;
+
+        public Usage(FileObject fileObject, int startOffset, int endOffset, String text) {
+            this.fileObject = fileObject;
+            this.startOffset = startOffset;
+            this.endOffset = endOffset;
+            this.text = text;
         }
-        return handle[0];
+
+        public FileObject getFileObject() {
+            return fileObject;
+        }
+
+        public int getStartOffset() {
+            return startOffset;
+        }
+
+        public int getEndOffset() {
+            return endOffset;
+        }
+
+        public String getText() {
+            return text;
+        }
     }
 }
