+++ enterprise/micronaut/src/org/netbeans/modules/micronaut/completion/MicronautConfigCompletionCollector.java
@@ -16,18 +16,19 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.modules.micronaut;
+package org.netbeans.modules.micronaut.completion;
 
-import java.util.Collections;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import javax.swing.text.Document;
 import org.netbeans.api.editor.document.EditorDocumentUtils;
 import org.netbeans.api.editor.mimelookup.MimeRegistration;
+import org.netbeans.api.lsp.Completion;
 import org.netbeans.api.project.FileOwnerQuery;
 import org.netbeans.api.project.Project;
 import org.netbeans.lib.editor.util.ArrayUtilities;
-import org.netbeans.spi.editor.completion.CompletionCollector;
+import org.netbeans.modules.micronaut.MicronautConfigProperties;
+import org.netbeans.spi.lsp.CompletionCollector;
 import org.openide.filesystems.FileObject;
 import org.springframework.boot.configurationmetadata.ConfigurationMetadataProperty;
 
@@ -39,42 +40,41 @@ import org.springframework.boot.configurationmetadata.ConfigurationMetadataPrope
 public class MicronautConfigCompletionCollector implements CompletionCollector {
 
     @Override
-    public boolean collectCompletions(Document doc, int offset, Context context, Consumer<Completion> consumer) {
+    public boolean collectCompletions(Document doc, int offset, Completion.Context context, Consumer<Completion> consumer) {
         FileObject fo = EditorDocumentUtils.getFileObject(doc);
         if (fo != null && "application.yml".equalsIgnoreCase(fo.getNameExt())) {
             Project project = FileOwnerQuery.getOwner(fo);
             if (project != null) {
-                MicronautConfigProperties configProperties = project.getLookup().lookup(MicronautConfigProperties.class);
-                if (configProperties != null) {
-                    new MicronautConfigCompletionTask().query(doc, offset, configProperties, new MicronautConfigCompletionTask.ItemFactory<CompletionCollector.Completion>() {
+                if (MicronautConfigProperties.hasConfigMetadata(project)) {
+                    new MicronautConfigCompletionTask().query(doc, offset, project, new MicronautConfigCompletionTask.ItemFactory<Completion>() {
                         @Override
-                        public CompletionCollector.Completion createTopLevelPropertyItem(String propName, int offset, int baseIndent, int indentLevelSize) {
+                        public Completion createTopLevelPropertyItem(String propName, int offset, int baseIndent, int indentLevelSize) {
                             StringBuilder insertText = new StringBuilder();
-                            TextFormat insertTextFormat = TextFormat.PlainText;
+                            Completion.TextFormat insertTextFormat = Completion.TextFormat.PlainText;
                             if ("*".equals(propName)) {
                                 insertText.append("$1:\n");
                                 ArrayUtilities.appendSpaces(insertText, baseIndent + indentLevelSize);
-                                insertTextFormat = TextFormat.Snippet;
+                                insertTextFormat = Completion.TextFormat.Snippet;
                             } else {
                                 insertText.append(propName).append(":\n");
                                 ArrayUtilities.appendSpaces(insertText, indentLevelSize);
                             }
-                            return CompletionCollector.Completion.newBuilder(propName).kind(Kind.Property).sortText(String.format("%4d%s", 10, propName))
+                            return CompletionCollector.newBuilder(propName).kind(Completion.Kind.Property).sortText(String.format("%4d%s", 10, propName))
                                     .insertText(insertText.toString()).insertTextFormat(insertTextFormat).build();
                         }
 
                         @Override
-                        public CompletionCollector.Completion createPropertyItem(ConfigurationMetadataProperty property, int offset, int baseIndent, int indentLevelSize, int idx) {
+                        public Completion createPropertyItem(ConfigurationMetadataProperty property, int offset, int baseIndent, int indentLevelSize, int idx) {
                             String[] parts = property.getId().substring(idx).split("\\.");
                             StringBuilder insertText = new StringBuilder();
                             int num = 1;
                             int indent = 0;
-                            TextFormat insertTextFormat = TextFormat.PlainText;
+                            Completion.TextFormat insertTextFormat = Completion.TextFormat.PlainText;
                             for (int i = 0; i < parts.length; i++) {
                                 String part = parts[i];
                                 if ("*".equals(part)) {
                                     insertText.append("$" + num++);
-                                    insertTextFormat = TextFormat.Snippet;
+                                    insertTextFormat = Completion.TextFormat.Snippet;
                                 } else {
                                     insertText.append(part);
                                 }
@@ -85,11 +85,11 @@ public class MicronautConfigCompletionCollector implements CompletionCollector {
                                     insertText.append(": ");
                                 }
                             }
-                            Completion.Builder builder = CompletionCollector.Completion.newBuilder(property.getId()).kind(Kind.Property)
+                            CompletionCollector.Builder builder = CompletionCollector.newBuilder(property.getId()).kind(Completion.Kind.Property)
                                     .sortText(String.format("%4d%s", property.isDeprecated() ? 30 : 20, property.getId())).insertText(insertText.toString())
                                     .insertTextFormat(insertTextFormat).documentation(CompletableFuture.completedFuture(new MicronautConfigDocumentation(property).getText()));
                             if (property.isDeprecated()) {
-                                builder.tags(Collections.singletonList(Tag.Deprecated));
+                                builder.addTag(Completion.Tag.Deprecated);
                             }
                             return builder.build();
                         }
