+++ enterprise/micronaut/src/org/netbeans/modules/micronaut/completion/MicronautConfigDocumentation.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.modules.micronaut;
+package org.netbeans.modules.micronaut.completion;
 
 import java.net.URL;
 import javax.swing.Action;
