+++ enterprise/micronaut/src/org/netbeans/modules/micronaut/resources/layer.xml
@@ -25,7 +25,7 @@
         <folder name="text">
             <folder name="x-java">
                 <folder name="AnnotationProcessors">
-                    <file name="org-netbeans-modules-micronaut-MicronautJavaConfigPropertiesCompletion.instance"/>
+                    <file name="org-netbeans-modules-micronaut-completion-MicronautJavaConfigPropertiesCompletion.instance"/>
                 </folder>
             </folder>
         </folder>
