+++ ide/api.lsp/build.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project basedir="." default="netbeans" name="ide/api.lsp">
+    <description>Builds, tests, and runs the project org.netbeans.api.lsp</description>
+    <import file="../../nbbuild/templates/projectized.xml"/>
+</project>
