+++ ide/api.lsp/manifest.mf
@@ -1,0 +1,6 @@
+Manifest-Version: 1.0
+OpenIDE-Module: org.netbeans.api.lsp/1
+OpenIDE-Module-Localizing-Bundle: org/netbeans/api/lsp/Bundle.properties
+OpenIDE-Module-Specification-Version: 1.0
+AutoUpdate-Show-In-Client: false
+
