+++ ide/api.lsp/nbproject/project.properties
@@ -1,0 +1,3 @@
+is.autoload=true
+javac.source=1.8
+javac.compilerargs=-Xlint -Xlint:-serial
