+++ ide/api.lsp/src/org/netbeans/api/lsp/Bundle.properties
@@ -14,9 +14,8 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-javac.compilerargs=-Xlint:unchecked
-javac.source=1.8
-javadoc.arch=${basedir}/arch.xml
-javadoc.apichanges=${basedir}/apichanges.xml
-spec.version.base=1.56.0
+OpenIDE-Module-Display-Category=Base IDE
+OpenIDE-Module-Long-Description=\
+    Provides APIs and support classes for Language Server implementations.
+OpenIDE-Module-Name=LSP APIs
+OpenIDE-Module-Short-Description=Language Server Protocol APIs
