+++ ide/editor.lib/apichanges.xml
@@ -83,21 +83,6 @@ is the proper place.
     <!-- ACTUAL CHANGES BEGIN HERE: -->
 
     <changes>
-        <change id="HyperlinkProviderExt.getHyperlinkLocation">
-            <summary>HyperlinkProviderExt.getHyperlinkLocation() method added</summary>
-            <version major="4" minor="20"/>
-            <date day="16" month="3" year="2021"/>
-            <author login="dbalek"/>
-            <compatibility binary="compatible" semantic="compatible" source="compatible" addition="yes" deprecation="no" deletion="no"/>
-            <description>
-                <p>
-                    <code>HyperlinkProviderExt.getHyperlinkLocation()</code> method added to allow for retrieving target location
-                    of a given hyperlink and sending it to outside NetBeans clients for display (e.g. via Language Server Protocol)
-                </p>
-            </description>
-            <class name="HyperlinkLocation" package="org.netbeans.lib.editor.hyperlink.spi"/>
-            <class name="HyperlinkProviderExt" package="org.netbeans.lib.editor.hyperlink.spi"/>
-        </change>
         <change id="caret-min-width">
             <summary>Allow to specify minimum caret width</summary>
             <version major="4" minor="17"/>
