+++ ide/editor.lib/nbproject/project.properties
@@ -17,7 +17,7 @@
 
 javac.compilerargs=-Xlint:unchecked
 javac.source=1.8
-spec.version.base=4.20.0
+spec.version.base=4.19.0
 is.autoload=true
 
 javadoc.arch=${basedir}/arch.xml
