+++ nbbuild/cluster.properties
@@ -311,6 +311,7 @@ nb.cluster.ide.depends=\
 nb.cluster.ide=\
         api.debugger,\
         api.java.classpath,\
+        api.lsp,\
         api.xml,\
         api.xml.ui,\
         bcpg,\
